morecommit
-updated onboarding layouts
-fixed minor bugs
-fixed bad format

+++ app/src/main/java/com/habitdev/sprout/ui/onBoarding/personalizationAssessment/adapter/Model/Result.java
@@ -27,8 +27,8 @@ public class Result {
         this.score = score;
     }
 
-    public int getTotal_count() {
-        return total_count;
+    public double getTotal_count() {
+        return (double) total_count;
     }
 
     public void setTotal_count(int total_count) {
